+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 -  `std::string_view` is now promoted to `std::string` to take a deep copy
 -  The queue capacity has been moved from `config` to `Tweakme.h`.
 -  Multiple formats patterns support for `stdout` and `stderr` handlers. See [example_stdout_multiple_formatters.cpp](https://github.com/odygrd/quill/blob/master/examples/example_custom_formatter.cpp)
+-  `quill::start()` will now block until the backend worker has started.
 -  Upgraded bundled libfmt to `6.2.1`
 
 ## v1.2.3